Reasonably complete generated README
